pyproject.toml
[project]
name = "hyper-strategies-pnl-weighted"
version = "0.1.0"
description = "Hyperliquid copytrading system with PnL-weighted dynamic allocation"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
backend = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "cachetools>=5.5",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "backend*"]
